requirements.txt
pygame~=2.0.1
neat-python~=0.92
graphviz~=0.17
matplotlib~=3.4.3
numpy~=1.21.2